Record new commit SHA. Permit non-numeric prefix. Fix termination.
Old calculation was printed out, but thrown away. Now, git reset --soft
is used to save the commit SHA.

The earlier "speedup" made a bug. Instead of comparing the iteration
count to the max, it compared the prefix to the count. This worked
if the prefix was numeric, but failed on hexic. This also meant
that the loop would never terminate.

[5361]